requirements.txt
accelerate
bitsandbytes 
sentence_transformers 
faiss-cpu 
langchain 
langchain_community
streamlit